+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Upcoming
 
+- [[#7](https://github.com/xsahil03x/super_paging/issues/7)] Fix generating load trigger notification when `itemCount` is less than `prefetchIndex`.
 - Added support for the `Pager.refresh` method to accept an optional `refreshKey` parameter.
 
 ## 0.1.0